Tell the user if they can't write the coffee cache
It's fairly easy to get into a situation wherein the user attempting to
run the coffeescript does not have permission to write the cache dir. We
don't want to change the behavior of the package too much, so let's just
log a message to stderr and continue, rather than exiting.